move PipelineLogger to dedicated logging package to avoid package cycles
