+++ src/main/java/com/atlassian/jira/cloud/jenkins/buildinfo/client/BuildsApi.java
@@ -6,12 +6,10 @@ import com.atlassian.jira.cloud.jenkins.common.client.ApiUpdateFailedException;
 import com.atlassian.jira.cloud.jenkins.common.client.JenkinsAppApi;
 import com.atlassian.jira.cloud.jenkins.common.client.JenkinsAppEventRequest;
 import com.atlassian.jira.cloud.jenkins.common.client.JenkinsAppRequest;
-import com.atlassian.jira.cloud.jenkins.util.PipelineLogger;
+import com.atlassian.jira.cloud.jenkins.logging.PipelineLogger;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import okhttp3.OkHttpClient;
 import org.jetbrains.annotations.NotNull;
-
-import java.io.PrintStream;
 
 public class BuildsApi extends JenkinsAppApi<BuildApiResponse> {
