+++ src/main/java/com/atlassian/jira/cloud/jenkins/buildinfo/freestyle/FreeStylePostBuildStep.java
@@ -9,7 +9,7 @@ import javax.annotation.Nullable;
 import javax.inject.Inject;
 
 import com.atlassian.jira.cloud.jenkins.config.JiraCloudSiteConfig;
-import com.atlassian.jira.cloud.jenkins.util.PipelineLogger;
+import com.atlassian.jira.cloud.jenkins.logging.PipelineLogger;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 import org.kohsuke.stapler.StaplerRequest;
